Main, login, post, register

+++ app/src/main/java/tk/ynrk/laporin/activity/LoginActivity.java
@@ -1,6 +1,8 @@
 package tk.ynrk.laporin.activity;
 
 import android.content.Intent;
+import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityOptionsCompat;
 import android.support.v4.util.Pair;
 import android.support.v7.app.AppCompatActivity;
@@ -11,6 +13,7 @@ import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -30,9 +33,11 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
     String BASE_URL;
 
     NetworkJSONLoader client;
-    private TextView appTitle, appDetail;
+    private TextView appTitle, appDetail, warning;
     private EditText email, password;
     private Button loginButton;
+    private CoordinatorLayout coordinatorLayout;
+    private LinearLayout redirect;
 
     private AppPreferences prefs;
 
@@ -47,6 +52,7 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
         email = findViewById(R.id.login_email_et);
         password = findViewById(R.id.login_password_et);
         loginButton = findViewById(R.id.login_button);
+        redirect = findViewById(R.id.redirect_to_register);
 
         BASE_URL = getResources().getString(R.string.BASE_URL);
         prefs = new AppPreferences(getApplicationContext());
@@ -64,15 +70,15 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
         findViewById(R.id.register_navigate_link).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-//                startActivity(new Intent(getApplicationContext(), RegisterActivity.class));
                 Intent intent = new Intent(getApplicationContext(), RegisterActivity.class);
                 Pair<View, String> p1 = Pair.create((View) appTitle, appTitle.getTransitionName());
                 Pair<View, String> p2 = Pair.create((View) appDetail, appDetail.getTransitionName());
                 Pair<View, String> p3 = Pair.create((View) email, email.getTransitionName());
                 Pair<View, String> p4 = Pair.create((View) password, password.getTransitionName());
                 Pair<View, String> p5 = Pair.create((View) loginButton, loginButton.getTransitionName());
+                Pair<View, String> p6 = Pair.create((View) redirect, redirect.getTransitionName());
                 ActivityOptionsCompat options = ActivityOptionsCompat.
-                        makeSceneTransitionAnimation(LoginActivity.this, p1, p2, p3, p4, p5);
+                        makeSceneTransitionAnimation(LoginActivity.this, p1, p2, p3, p4, p5, p6);
                 startActivity(intent, options.toBundle());
             }
         });
@@ -80,12 +86,18 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
         loginButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                if (!checkForm()){
+                    return;
+                }
+
                 hideKeyboard(view);
                 String token = prefs.getValue("token");
                 Log.d("LOGIN_ACTIVITY", "token " + token);
+
                 if (token != null || !token.isEmpty()) {
                     prefs.clear();
                 }
+
                 Map<String, Object> data = new HashMap<>();
                 data.put("email", email.getText().toString());
                 data.put("password", password.getText().toString());
@@ -93,6 +105,15 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
             }
         });
 
+    }
+
+    public boolean checkForm() {
+        if(email.getText().toString().isEmpty() || password.getText().toString().isEmpty()){
+            Toast.makeText(this, "Form tidak boleh kosong", Toast.LENGTH_LONG).show();
+            return false;
+        } else {
+            return true;
+        }
     }
 
     public void hideKeyboard(View v) {
@@ -112,6 +133,11 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
             Toast.makeText(this, "DONE", Toast.LENGTH_SHORT).show();
             Log.d("LOGIN_ACTIVITY", token);
             prefs.setValue("token", token);
+
+            //Navigate to main screen after login in
+            Intent intent = new Intent(getApplicationContext(), MainActivity.class);
+            startActivity(intent);
+            finish();
         } catch (JSONException err) {
             Log.e("LOGIN_ACTIVITY", err.toString());
         }
@@ -119,8 +145,8 @@ public class LoginActivity extends AppCompatActivity implements NetworkJSONLoade
 
     @Override
     public void onTaskError(VolleyError error) {
-        String errorString = new String(error.networkResponse.data);
-        Log.e("LOGIN_ACTIVITY", "" + errorString);
+        Log.e("LOGIN_ACTIVITY", "" + "err");
         Toast.makeText(this, "ERROR", Toast.LENGTH_SHORT).show();
+        Snackbar.make(findViewById(R.id.login_parent), "Login failed please try again", Snackbar.LENGTH_SHORT);
     }
 }
